feat: html to epub - refactor library (#12)
* chore: revert install library

* refactor: extract constants

* fix: import type instead of element from deprecated library

FIXME: replace hast for rehype

* refactor: extract types (no changes)

* refactor: extract function (no changes

* fix: adapt html-to-epub tests

* chore: delete mochaand asset to use jest instead.

* chore: use default __dirname

* chore: update jest. ignore test result files

* chore: condense tests

* refactor: move local function to class field. unifiy if conditions

* chore: restore tests

* fix: use safe uuid generator

* chore: fix import

* chore: change jest version for compatibility

* chore: rename disabled tests
