+++ package.json
@@ -16,7 +16,6 @@
     "ejs": "^3.1.9",
     "entities": "^4.5.0",
     "fs-extra": "^11.2.0",
-    "hast": "^1.0.0",
     "image-size": "^1.1.1",
     "jszip": "2.5.0",
     "mime": "^4.0.1",
@@ -35,7 +34,6 @@
     "@types/ejs": "^3.1.5",
     "@types/fs-extra": "^11.0.4",
     "@types/jest": "^29.5.12",
-    "@types/mocha": "^10.0.6",
     "@types/node": "^20.11.20",
     "@types/request": "^2.48.12",
     "@types/uslug": "^1.0.4",
@@ -52,7 +50,6 @@
     "eslint-plugin-prettier": "^4.2.1",
     "jest": "^28.0.0",
     "jest-junit": "^14.0.0",
-    "mocha": "^10.3.0",
     "prettier": "^2.3.0",
     "ts-jest": "^28.0.0",
     "ts-node": "^10.4.0",
