+++ src/constants-gen.ts
@@ -1,0 +1,217 @@
+// Allowed HTML attributes & tags
+export const allowedAttributes = [
+  "content",
+  "alt",
+  "id",
+  "title",
+  "src",
+  "href",
+  "about",
+  "accesskey",
+  "aria-activedescendant",
+  "aria-atomic",
+  "aria-autocomplete",
+  "aria-busy",
+  "aria-checked",
+  "aria-controls",
+  "aria-describedat",
+  "aria-describedby",
+  "aria-disabled",
+  "aria-dropeffect",
+  "aria-expanded",
+  "aria-flowto",
+  "aria-grabbed",
+  "aria-haspopup",
+  "aria-hidden",
+  "aria-invalid",
+  "aria-label",
+  "aria-labelledby",
+  "aria-level",
+  "aria-live",
+  "aria-multiline",
+  "aria-multiselectable",
+  "aria-orientation",
+  "aria-owns",
+  "aria-posinset",
+  "aria-pressed",
+  "aria-readonly",
+  "aria-relevant",
+  "aria-required",
+  "aria-selected",
+  "aria-setsize",
+  "aria-sort",
+  "aria-valuemax",
+  "aria-valuemin",
+  "aria-valuenow",
+  "aria-valuetext",
+  "className",
+  "content",
+  "contenteditable",
+  "contextmenu",
+  "datatype",
+  "dir",
+  "draggable",
+  "dropzone",
+  "hidden",
+  "hreflang",
+  "id",
+  "inlist",
+  "itemid",
+  "itemref",
+  "itemscope",
+  "itemtype",
+  "lang",
+  "media",
+  "ns1:type",
+  "ns2:alphabet",
+  "ns2:ph",
+  "onabort",
+  "onblur",
+  "oncanplay",
+  "oncanplaythrough",
+  "onchange",
+  "onclick",
+  "oncontextmenu",
+  "ondblclick",
+  "ondrag",
+  "ondragend",
+  "ondragenter",
+  "ondragleave",
+  "ondragover",
+  "ondragstart",
+  "ondrop",
+  "ondurationchange",
+  "onemptied",
+  "onended",
+  "onerror",
+  "onfocus",
+  "oninput",
+  "oninvalid",
+  "onkeydown",
+  "onkeypress",
+  "onkeyup",
+  "onload",
+  "onloadeddata",
+  "onloadedmetadata",
+  "onloadstart",
+  "onmousedown",
+  "onmousemove",
+  "onmouseout",
+  "onmouseover",
+  "onmouseup",
+  "onmousewheel",
+  "onpause",
+  "onplay",
+  "onplaying",
+  "onprogress",
+  "onratechange",
+  "onreadystatechange",
+  "onreset",
+  "onscroll",
+  "onseeked",
+  "onseeking",
+  "onselect",
+  "onshow",
+  "onstalled",
+  "onsubmit",
+  "onsuspend",
+  "ontimeupdate",
+  "onvolumechange",
+  "onwaiting",
+  "prefix",
+  "property",
+  "rel",
+  "resource",
+  "rev",
+  "role",
+  "spellcheck",
+  "style",
+  "tabindex",
+  "target",
+  "title",
+  "type",
+  "typeof",
+  "vocab",
+  "xml:base",
+  "xml:lang",
+  "xml:space",
+  "colspan",
+  "rowspan",
+  "epub:type",
+  "epub:prefix",
+];
+export const allowedXhtml11Tags = [
+  "div",
+  "p",
+  "h1",
+  "h2",
+  "h3",
+  "h4",
+  "h5",
+  "h6",
+  "ul",
+  "ol",
+  "li",
+  "dl",
+  "dt",
+  "dd",
+  "address",
+  "hr",
+  "pre",
+  "blockquote",
+  "center",
+  "ins",
+  "del",
+  "a",
+  "span",
+  "bdo",
+  "br",
+  "em",
+  "strong",
+  "dfn",
+  "code",
+  "samp",
+  "kbd",
+  "bar",
+  "cite",
+  "abbr",
+  "acronym",
+  "q",
+  "sub",
+  "sup",
+  "tt",
+  "i",
+  "b",
+  "big",
+  "small",
+  "u",
+  "s",
+  "strike",
+  "basefont",
+  "font",
+  "object",
+  "param",
+  "img",
+  "table",
+  "caption",
+  "colgroup",
+  "col",
+  "thead",
+  "tfoot",
+  "tbody",
+  "tr",
+  "th",
+  "td",
+  "embed",
+  "applet",
+  "iframe",
+  "img",
+  "map",
+  "noscript",
+  "ns:svg",
+  "object",
+  "script",
+  "table",
+  "tt",
+  "var",
+];
